requirements.txt
python-dotenv==1.0.1
Requests==2.32.3
spotipy==2.24.0
